move history to own module
