+++ pypesto/history/generate.py
@@ -2,16 +2,18 @@ from pathlib import Path
 from typing import Sequence
 
 from ..C import SUFFIXES_CSV, SUFFIXES_HDF5
-from .base import History, HistoryBase, HistoryTypeError
+from .base import History, HistoryBase
 from .csv import CsvHistory
 from .hdf5 import Hdf5History
 from .memory import MemoryHistory
+from .options import HistoryOptions
+from .util import HistoryTypeError
 
 
-def create_history_from_options(
-    self,
+def create_history(
     id: str,
     x_names: Sequence[str],
+    options: HistoryOptions,
 ) -> HistoryBase:
     """Create a :class:`HistoryBase` object; Factory method.
 
@@ -21,24 +23,31 @@ def create_history_from_options(
         Identifier for the history.
     x_names:
         Parameter names.
+    options:
+        History options.
+
+    Returns
+    -------
+    history:
+        A history object corresponding to the inputs.
     """
     # create different history types based on the inputs
-    if self.storage_file is None:
-        if self.trace_record:
-            return MemoryHistory(options=self)
+    if options.storage_file is None:
+        if options.trace_record:
+            return MemoryHistory(options=options)
         else:
-            return History(options=self)
+            return History(options=options)
 
     # replace id template in storage file
-    storage_file = self.storage_file.replace("{id}", id)
+    storage_file = options.storage_file.replace("{id}", id)
 
     # evaluate type
     suffix = Path(storage_file).suffix[1:]
 
     # create history type based on storage type
     if suffix in SUFFIXES_CSV:
-        return CsvHistory(x_names=x_names, file=storage_file, options=self)
+        return CsvHistory(x_names=x_names, file=storage_file, options=options)
     elif suffix in SUFFIXES_HDF5:
-        return Hdf5History(id=id, file=storage_file, options=self)
+        return Hdf5History(id=id, file=storage_file, options=options)
     else:
         raise HistoryTypeError(suffix)
