+++ pypesto/objective/__init__.py
@@ -14,4 +14,3 @@ from .priors import (
     NegLogPriors,
     get_parameter_prior_dict,
 )
-from .util import res_to_chi2, sres_to_schi2
