+++ pypesto/sample/metropolis.py
@@ -3,7 +3,8 @@ from typing import Dict, Sequence, Union
 import numpy as np
 from tqdm import tqdm
 
-from ..objective import History, NegLogPriors, ObjectiveBase
+from ..history import History
+from ..objective import NegLogPriors, ObjectiveBase
 from ..problem import Problem
 from ..result import McmcPtResult
 from .sampler import InternalSample, InternalSampler
